+++ lib/membrane/core/bin.ex
@@ -82,12 +82,13 @@ defmodule Membrane.Core.Bin do
 
   @impl GenServer
   def init(options) do
-    Process.monitor(options.parent)
+    %{parent: parent, name: name, module: module, log_metadata: log_metadata} = options
 
-    %{name: name, module: module, log_metadata: log_metadata} = options
+    Process.monitor(parent)
+
     name_str = if String.valid?(name), do: name, else: inspect(name)
     :ok = Membrane.Logger.set_prefix(name_str <> " bin")
-    Logger.metadata(log_metadata)
+    :ok = Logger.metadata(log_metadata)
     :ok = ComponentPath.set_and_append(log_metadata[:parent_path] || [], name_str <> " bin")
 
     Telemetry.report_init(:bin)
@@ -118,7 +119,7 @@ defmodule Membrane.Core.Bin do
            CallbackHandler.exec_and_handle_callback(
              :handle_init,
              Membrane.Core.Bin.ActionHandler,
-             %{state: false},
+             %{},
              [options.user_options],
              state
            ) do
@@ -153,13 +154,11 @@ defmodule Membrane.Core.Bin do
 
   @impl GenServer
   def handle_call(
-        Message.new(:handle_link, [direction, this, other, other_info, metadata]),
+        Message.new(:handle_link, [direction, this, other, params]),
         _from,
         state
       ) do
-    {reply, state} =
-      PadController.handle_link(direction, this, other, other_info, metadata, state)
-
+    {reply, state} = PadController.handle_link(direction, this, other, params, state)
     {:reply, reply, state}
   end