+++ lib/membrane/core/bin/pad_controller.ex
@@ -19,7 +19,7 @@ defmodule Membrane.Core.Bin.PadController do
   require Membrane.Pad
 
   @doc """
-  Handles link request from outside of the bin.
+  Handles a link request from the bin's parent.
   """
   @spec handle_external_link_request(
           Pad.ref_t(),
@@ -46,38 +46,31 @@ defmodule Membrane.Core.Bin.PadController do
     pad_options = Child.PadController.parse_pad_options!(pad_name, pad_options, state)
 
     state =
-      with {:ok, response_received?} <- PadModel.get_data(state, pad_ref, :response_received?) do
-        if response_received? do
-          Membrane.Logger.debug("Sending link response, #{inspect(pad_ref)}")
-          Message.send(state.parent_pid, :link_response, link_id)
-        end
-
-        state
-      else
+      case PadModel.get_data(state, pad_ref) do
         {:error, {:unknown_pad, pad_ref}} ->
           init_pad_data(pad_ref, info, state)
+
+        # This case is for pads that were instantiated before the external link request,
+        # that is in the internal link request (see `handle_internal_link_request/4`).
+        # This is possible only for static pads. It might have happened that we already
+        # received a link response for such a pad, so we should reply immediately.
+        {:ok, data} ->
+          if data.response_received? do
+            Membrane.Logger.debug("Sending link response, #{inspect(pad_ref)}")
+            Message.send(state.parent_pid, :link_response, link_id)
+          end
+
+          state
       end
 
-    {spec_ref, state} =
-      PadModel.get_and_update_data!(
-        state,
-        pad_ref,
-        &{&1.spec_ref, %{&1 | link_id: link_id, options: pad_options}}
-      )
-
-    state =
-      if spec_ref do
-        ChildLifeController.LinkHandler.proceed_spec_linking(spec_ref, state)
-      else
-        state
-      end
-
-    {:ok, state} = handle_pad_added(pad_ref, state)
+    state = PadModel.update_data!(state, pad_ref, &%{&1 | link_id: link_id, options: pad_options})
+    {:ok, state} = maybe_handle_pad_added(pad_ref, state)
     state
   end
 
   @doc """
-  Handles link request from one of bin's children.
+  Handles a link request coming from the bin itself when linking a bin's pad
+  to a pad of one of its children.
   """
   @spec handle_internal_link_request(
           Pad.ref_t(),
@@ -85,7 +78,7 @@ defmodule Membrane.Core.Bin.PadController do
           ChildLifeController.spec_ref_t(),
           State.t()
         ) :: State.t()
-  def handle_internal_link_request(pad_ref, endpoint, spec_ref, state) do
+  def handle_internal_link_request(pad_ref, child_endpoint, spec_ref, state) do
     pad_name = Pad.name_by_ref(pad_ref)
     info = Map.fetch!(state.pads.info, pad_name)
 
@@ -94,6 +87,7 @@ defmodule Membrane.Core.Bin.PadController do
         :ok == PadModel.assert_instance(state, pad_ref) ->
           state
 
+        # Static pads can be linked internally before the external link request
         info.availability == :always ->
           init_pad_data(pad_ref, info, state)
 
@@ -101,7 +95,11 @@ defmodule Membrane.Core.Bin.PadController do
           raise LinkError, "Dynamic pads must be firstly linked externally, then internally"
       end
 
-    PadModel.update_data!(state, pad_ref, &%{&1 | endpoint: endpoint, spec_ref: spec_ref})
+    PadModel.update_data!(
+      state,
+      pad_ref,
+      &%{&1 | endpoint: child_endpoint, spec_ref: spec_ref}
+    )
   end
 
   @doc """
@@ -145,21 +143,23 @@ defmodule Membrane.Core.Bin.PadController do
           Pad.direction_t(),
           LinkParser.raw_endpoint_t(),
           LinkParser.raw_endpoint_t(),
-          PadModel.pad_info_t() | nil,
-          map,
+          %{initiator: :parent}
+          | %{initiator: :sibling, other_info: PadModel.pad_info_t() | nil, link_metadata: map},
           Core.Bin.State.t()
         ) :: Type.stateful_try_t(PadModel.pad_info_t(), Core.Bin.State.t())
-  def handle_link(direction, this, other, other_info, link_metadata, state) do
-    pad_data = PadModel.get_data!(state, this.pad_ref)
-    %{spec_ref: spec_ref, endpoint: endpoint} = pad_data
+  def handle_link(direction, endpoint, other_endpoint, params, state) do
+    pad_data = PadModel.get_data!(state, endpoint.pad_ref)
+    %{spec_ref: spec_ref, endpoint: child_endpoint} = pad_data
 
     pad_props =
-      Map.merge(this.pad_props, endpoint.pad_props, fn key, external_value, internal_value ->
+      Map.merge(endpoint.pad_props, child_endpoint.pad_props, fn key,
+                                                                 external_value,
+                                                                 internal_value ->
         if key in [
-             :demand_excess_factor,
+             :demand_excess,
              :min_demand_factor,
-             :auto_demand_size_factor,
-             :toilet_capacity_factor
+             :auto_demand_size,
+             :toilet_capacity
            ] do
           external_value || internal_value
         else
@@ -167,42 +167,43 @@ defmodule Membrane.Core.Bin.PadController do
         end
       end)
 
-    endpoint = %{endpoint | pad_props: pad_props}
+    child_endpoint = %{child_endpoint | pad_props: pad_props}
 
-    :ok =
-      Child.PadController.validate_pad_mode!(
-        {this.pad_ref, pad_data},
-        {other.pad_ref, other_info}
-      )
+    if params.initiator == :sibling do
+      :ok =
+        Child.PadController.validate_pad_mode!(
+          {endpoint.pad_ref, pad_data},
+          {other_endpoint.pad_ref, params.other_info}
+        )
+    end
 
     reply =
-      Message.call(endpoint.pid, :handle_link, [
+      Message.call(child_endpoint.pid, :handle_link, [
         direction,
-        endpoint,
-        other,
-        other_info,
-        link_metadata
+        child_endpoint,
+        other_endpoint,
+        params
       ])
 
-    state = PadModel.set_data!(state, this.pad_ref, :linked?, true)
-    state = PadModel.set_data!(state, this.pad_ref, :endpoint, endpoint)
+    state = PadModel.set_data!(state, endpoint.pad_ref, :linked?, true)
+    state = PadModel.set_data!(state, endpoint.pad_ref, :endpoint, child_endpoint)
     state = ChildLifeController.LinkHandler.proceed_spec_linking(spec_ref, state)
     {reply, state}
   end
 
   @doc """
-  Handles situation where pad has been unlinked (e.g. when connected element has been removed from pipline)
+  Handles situation where the pad has been unlinked (e.g. when connected element has been removed from the pipeline)
   """
   @spec handle_unlink(Pad.ref_t(), Core.Bin.State.t()) :: Type.stateful_try_t(Core.Bin.State.t())
   def handle_unlink(pad_ref, state) do
-    with {:ok, state} <- handle_pad_removed(pad_ref, state) do
+    with {:ok, state} <- maybe_handle_pad_removed(pad_ref, state) do
       PadModel.delete_data(state, pad_ref)
     end
   end
 
-  @spec handle_pad_added(Pad.ref_t(), Core.Bin.State.t()) ::
+  @spec maybe_handle_pad_added(Pad.ref_t(), Core.Bin.State.t()) ::
           Type.stateful_try_t(Core.Bin.State.t())
-  defp handle_pad_added(ref, state) do
+  defp maybe_handle_pad_added(ref, state) do
     %{options: pad_opts, direction: direction, availability: availability} =
       PadModel.get_data!(state, ref)
 
@@ -221,9 +222,9 @@ defmodule Membrane.Core.Bin.PadController do
     end
   end
 
-  @spec handle_pad_removed(Pad.ref_t(), Core.Bin.State.t()) ::
+  @spec maybe_handle_pad_removed(Pad.ref_t(), Core.Bin.State.t()) ::
           Type.stateful_try_t(Core.Bin.State.t())
-  defp handle_pad_removed(ref, state) do
+  defp maybe_handle_pad_removed(ref, state) do
     %{direction: direction, availability: availability} = PadModel.get_data!(state, ref)
 
     if Pad.availability_mode(availability) == :dynamic do