+++ lib/membrane/core/element/toilet.ex
@@ -1,7 +1,7 @@
 defmodule Membrane.Core.Element.Toilet do
   @moduledoc false
 
-  # Toilet is an entity that can be urinated to and rinsed. If it's not rinsed on
+  # Toilet is an entity that can be filled and drained. If it's not drained on
   # time and exceeds its capacity, it overflows by logging an error and killing
   # the responsible process (passed on the toilet creation).
 
@@ -9,28 +9,27 @@ defmodule Membrane.Core.Element.Toilet do
 
   @opaque t :: {__MODULE__, :atomics.atomics_ref(), pos_integer, Process.dest()}
 
-  @spec default_capacity_factor() :: number
-  def default_capacity_factor, do: 200
+  @default_capacity_factor 200
 
   @spec new(pos_integer() | nil, Membrane.Buffer.Metric.unit_t(), Process.dest()) :: t
-  def new(capacity_factor, demand_unit, responsible_process) do
-    capacity =
+  def new(capacity, demand_unit, responsible_process) do
+    default_capacity =
       Membrane.Buffer.Metric.from_unit(demand_unit).buffer_size_approximation() *
-        (capacity_factor || default_capacity_factor())
+        @default_capacity_factor
 
-    capacity = ceil(capacity)
+    capacity = capacity || default_capacity
     {__MODULE__, :atomics.new(1, []), capacity, responsible_process}
   end
 
-  @spec urinate(t, non_neg_integer) :: :ok
-  def urinate({__MODULE__, atomic, capacity, responsible_process}, amount) do
+  @spec fill(t, non_neg_integer) :: :ok
+  def fill({__MODULE__, atomic, capacity, responsible_process}, amount) do
     size = :atomics.add_get(atomic, 1, amount)
     if size > capacity, do: overflow(size, capacity, responsible_process)
     :ok
   end
 
-  @spec rinse(t, non_neg_integer) :: :ok
-  def rinse({__MODULE__, atomic, _capacity, _responsible_process}, amount) do
+  @spec drain(t, non_neg_integer) :: :ok
+  def drain({__MODULE__, atomic, _capacity, _responsible_process}, amount) do
     :atomics.sub(atomic, 1, amount)
   end
 
@@ -69,7 +68,7 @@ defmodule Membrane.Core.Element.Toilet do
     processes the stream too slow or doesn't process it at all.
     To have control over amount of buffers being produced, consider using output in pull mode
     (see `Membrane.Pad.mode_t`).
-    You can also try changing the `toilet_capacity_factor` in `Membrane.ParentSpec.via_in/3`.
+    You can also try changing the `toilet_capacity` in `Membrane.ParentSpec.via_in/3`.
     """)
 
     Process.exit(responsible_process, :kill)