+++ test/membrane/core/element/event_controller_test.exs
@@ -30,7 +30,7 @@ defmodule Membrane.Core.Element.EventControllerTest do
         demand_pad: :some_pad,
         log_tag: "test",
         toilet?: false,
-        demand_excess_factor: nil,
+        demand_excess: nil,
         min_demand_factor: nil
       })